requirement.txt
discord.py
python-dotenv
Flask
requests